+++ src/nvmain/entrypoint.sh
@@ -3,5 +3,6 @@
 echo "[NVMSIM] entrypoint.sh -> starting NVMAIN Docker entrypoint"
 cd /opt/nvmain;
 # scons --build-type=fast
-./nvmain.fast Config/2D_DRAM_example.config tracefile.nvt 0
+./nvmain.fast Config/2D_DRAM_example.config tracefile.nvt 1000 TraceReader=SniperTrace
+# ./nvmain.fast 
 echo "[NVMSIM] entrypoint.sh -> finishing NVMAIN Docker entrypoint"