+++ src/sniper/entrypoint.sh
@@ -2,5 +2,5 @@
 
 echo "[NVMSIM] entrypoint.sh -> starting SNIPER Docker entrypoint"
 cd /opt/sniper-7.2;
-/opt/sniper-7.2/run-sniper -d /nvmsim/logs -- /bin/date
+/opt/sniper-7.2/run-sniper -d /nvmsim/logs -- touch /tmp/test.txt
 echo "[NVMSIM] entrypoint.sh -> finishing SNIPER Docker entrypoint"